CE - Multi-Agent Orchesration/api/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlmodel>=0.0.14
sse-starlette>=1.8.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0
